Finish CIRCLE, fix STR$, fix scrolling
STR$ must return a leading space also when printing an integer.
Some formatting of graps relies on this and would remove the first digit
of a mult-digit number. Added test.
In tkinter text-mode the scrolling wasn't working, this is now fixed.
Text-mode in tkinter didn't correctly the dimension of the text widget.
We now alternate between text and graphics display by unmapping the one
currently unused. This is because both do not fit on the screen.
In cmd_input (implementing INPUT) we need to add a newline when the
users presses the key. This has no effect in the middle of the screen
(except for changing cursor position) but is essential when at the end
of the screen where we need to scroll. We now hopefully correctly
implement a single carriage-return without a newline.
Fix CSRLIN off-by-one error.