More fixes for MBF float
Now we employ the sticky-bit shifting used before add/sub.
Also fix rounding again.
This now seems to reproduce the asm implementation (for the testcases we
have). Both addition and multiplication can have 1 bit difference to
single precision. For addition the reason is unclear. For multiplication
it happens because the low-bits are thrown away (instead of computing
the sticky bit).